chore: even moar simplify
